+++ platform/RELEASES.md
@@ -107,3 +107,13 @@ Based on:
 - [csharp v3.2.0] platform
 ### Releases
 - [NuGet v3.2.0] https://www.nuget.org/packages/Codat.Platform/3.2.0 - platform
+
+## 2023-12-21 15:16:24
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.128.1 (2.223.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.3.0] platform
+### Releases
+- [NuGet v3.3.0] https://www.nuget.org/packages/Codat.Platform/3.3.0 - platform